third_party/blink/web_tests/svg/dynamic-updates/SVGFEColorMatrixElement-dom-values-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'values' attribute of the SVGFEColorMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS matrixElement.getAttribute('values') is ".45 .45 .5 0 0 .13 .13 .13 0 0 .33 .13 .33 0 0 .99 .56 .33 0 0"
PASS saturateElement.getAttribute('values') is "0.5"
PASS hueRotateElement.getAttribute('values') is "10"
PASS matrixElement.getAttribute('values') is ".33 .33 .33 0 0 .33 .33 .33 0 0 .33 .33 .33 0 0 .33 .33 .33 0 0"
PASS saturateElement.getAttribute('values') is "0.25"
PASS hueRotateElement.getAttribute('values') is "90"
PASS successfullyParsed is true

TEST COMPLETE
